Make some fixtures shared between all tests
The Elasticsearch client will be needed for tests of the search
results, so it's easier if it's accessible to all test files. Same
for the test glossary entries.
